+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Extended **Validation API**: you can define your own predefined validators via `.define_validator(name, &validation)` directive;
+- `re_setting` - a special DSL command method that fully redefines existing settings (redefines existing settings instead of reopening them);
 
 ## [0.19.1] - 2019-11-29
 ### Changed